requirements.txt
asgiref==3.7.2
colorama==0.4.6
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-crispy-forms==2.1
django-import-export==3.3.7
django-pandas==0.6.6
django-registration-redux==2.13
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
plotly==5.19.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
